Renaming ItemDetailContainer directory and it's imports
