+++ src/components/itemDetailContainer/itemDetailContainer.scss
