+++ src/pages/home/homePage.js
@@ -1,7 +1,7 @@
 import './homePage.scss';
 import { NavBar } from '../../components/navBar/navBar';
 import { ItemListContainer } from '../../components/itemListContainer/itemListContainer';
-import { ItemDetailContainer } from '../../components/itemDetailContainer copy/itemDetailContainer';
+import { ItemDetailContainer } from '../../components/itemDetailContainer/itemDetailContainer';
 import { WORDINGS } from '../../wordings';
 
 export const HomePage = () => {
